externalize plugins package
