+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/config/crd/crd.go
