+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/config/crd/kustomization.go
