+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/config/rbac/role.go
