+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/config/testing/debug_logs_patch.go
