+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/config/testing/kustomization.go
