+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/config/testing/manager_image.go
