+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/gitignore.go
