+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/makefile.go
