+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/molecule/mdefault/create.go
