+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/molecule/mdefault/prepare.go
