+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/molecule/mdefault/tasks_test_resource.go
