+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/molecule/mkind/create.go
