+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/playbooks/placeholder.go
