+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/roles/placeholder.go
