+++ pkg/plugins/ansible/v1/scaffolds/internal/templates/roles/templates_dir.go
@@ -19,27 +19,25 @@ import (
 
 	"sigs.k8s.io/kubebuilder/v3/pkg/machinery"
 
-	"github.com/operator-framework/ansible-operator-plugins/internal/plugins/ansible/v1/constants"
+	"github.com/operator-framework/ansible-operator-plugins/pkg/plugins/ansible/v1/constants"
 )
 
-var _ machinery.Template = &HandlersMain{}
+var _ machinery.Template = &RoleTemplates{}
 
-type HandlersMain struct {
+type RoleTemplates struct {
 	machinery.TemplateMixin
 	machinery.ResourceMixin
 }
 
 // SetTemplateDefaults implements machinery.Template
-func (f *HandlersMain) SetTemplateDefaults() error {
+func (f *RoleTemplates) SetTemplateDefaults() error {
 	if f.Path == "" {
-		f.Path = filepath.Join(constants.RolesDir, "%[kind]", "handlers", "main.yml")
-		f.Path = f.Resource.Replacer().Replace(f.Path)
+		f.Path = filepath.Join(constants.RolesDir, "%[kind]", "templates", ".placeholder")
 	}
+	f.Path = f.Resource.Replacer().Replace(f.Path)
 
-	f.TemplateBody = handlersMainAnsibleTmpl
+	f.TemplateBody = templatesDirPlaceholder
 	return nil
 }
 
-const handlersMainAnsibleTmpl = `---
-# handlers file for {{ .Resource.Kind }}
-`
+const templatesDirPlaceholder = ``
