+++ pkg/plugins/util/cleanup.go
